chore: renamed AddIncludesToMarkdown
