+++ src/plugins/AddIncludesToMarkdown.js
